Move Datasets class to distributed.publish module
